+++ src/main/java/com/aicodinator/backend/domain/region/repository/RegionRepository.java
@@ -1,6 +1,6 @@
 package com.aicodinator.backend.domain.region.repository;
 
-import com.aicodinator.backend.domain.region.domain.Region;
+import com.aicodinator.backend.domain.region.domain.entity.Region;
 import com.aicodinator.backend.domain.user.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
